* src/SDCCicode.c (geniCodeAddressOf): fixed part of bug #840381 * src/SDCCopt.c (replaceRegEqv, isLocalWithoutDef): fixed other part (SIGSEV) of bug #840381 * src/SDCCmain.c (linkEdit, assemble): fixed bug #841606 (don't unlink new file before rename if new and old filenames are the same)
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@3015 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ src/SDCCmain.c
@@ -1616,7 +1616,8 @@ linkEdit (char **envp)
       strncatz (scratchFileName,
 		options.out_fmt ? ".S19" : ".ihx",
 		sizeof(scratchFileName));
-      unlink (fullDstFileName);
+      if (strcmp (fullDstFileName, scratchFileName))
+	unlink (fullDstFileName);
       rename (scratchFileName, fullDstFileName);
 
       strncpyz (buffer, fullDstFileName, sizeof(buffer));
@@ -1631,13 +1632,15 @@ linkEdit (char **envp)
       strncatz (scratchFileName, ".map", sizeof(scratchFileName));
       *q = 0;
       strncatz(buffer, ".map", sizeof(buffer));
-      unlink (buffer);
+      if (strcmp (scratchFileName, buffer))
+	unlink (buffer);
       rename (scratchFileName, buffer);
       *p = 0;
       strncatz (scratchFileName, ".mem", sizeof(scratchFileName));
       *q = 0;
       strncatz(buffer, ".mem", sizeof(buffer));
-      unlink (buffer);
+      if (strcmp (scratchFileName, buffer))
+        unlink (buffer);
       rename (scratchFileName, buffer);
       if (options.debug)
         {
@@ -1645,12 +1648,14 @@ linkEdit (char **envp)
           strncatz (scratchFileName, ".cdb", sizeof(scratchFileName));
           *q = 0;
           strncatz(buffer, ".cdb", sizeof(buffer));
-          unlink (buffer);
+          if (strcmp (scratchFileName, buffer))
+            unlink (buffer);
           rename (scratchFileName, buffer);
           /* and the OMF file without extension: */
           *p = 0;
           *q = 0;
-          unlink (buffer);
+          if (strcmp (scratchFileName, buffer))
+            unlink (buffer);
           rename (scratchFileName, buffer);
         }
     }
@@ -1702,7 +1707,8 @@ assemble (char **envp)
         strncatz (scratchFileName, 
 		  port->linker.rel_ext,
 		  sizeof(scratchFileName));
-        unlink (fullDstFileName);
+        if (strcmp (scratchFileName, fullDstFileName))
+          unlink (fullDstFileName);
         rename (scratchFileName, fullDstFileName);
     }
 }
